backend_fastapi/requirements.txt
fastapi
uvicorn
torch
torchvision
torchaudio
opencv-python
pillow
numpy
python-multipart
realesrgan
basicsr
ultralytics
